Refactored output_file_handle to also use the AbstractFileHandle class
